account for dispatch_tasks

+++ app/jobs/nightly_syncs_job.rb
@@ -29,12 +29,13 @@ class NightlySyncsJob < CaseflowJob
       next if legacy_appeal.case_record.present? # extra check
 
       # delete pure danglers
-      if legacy_appeal.tasks.none?
+      if legacy_appeal.tasks.none? && legacy_appeal.dispatch_tasks.none?
         legacy_appeal.destroy!
       else
         # if we have tasks and no case_record, then we need to cancel all the tasks,
         # but we do not delete the dangling LegacyAppeal record.
         legacy_appeal.tasks.open.where(parent_id: nil).each(&:cancel_task_and_child_subtasks)
+        legacy_appeal.dispatch_tasks.open.each(&:invalidate!)
       end
     end
     datadog_report_time_segment(segment: "sync_cases_from_vacols", start_time: start_time)
